clean code

+++ src/Controller/MediaController.php
@@ -39,8 +39,6 @@ class MediaController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $image = $form->get('image')->getData();
 
-            // this condition is needed because the 'brochure' field is not required
-            // so the PDF file must be processed only when a file is uploaded
             if ($image) {
                 $result =array();
                 foreach ($image as $image)
@@ -97,12 +95,8 @@ class MediaController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-
             $image = $form->get('mediaImage')->getData();
-           // $username = $form["user"]["username"]->getData();
 
-            // this condition is needed because the 'brochure' field is not required
-            // so the PDF file must be processed only when a file is uploaded
             foreach ($image as $image) {
                 $originalFilename = pathinfo($image->getClientOriginalName(), PATHINFO_FILENAME);
                 // this is needed to safely include the file name as part of the URL
@@ -125,7 +119,7 @@ class MediaController extends AbstractController
             }
             $mediaRepository->add($media, true);
 
-            return $this->redirectToRoute('app_media_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_tricks_index', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('media/edit.html.twig', [
@@ -135,14 +129,14 @@ class MediaController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/delete", name="app_media_delete", methods={"GET"})
+     * @Route("/{id}/delete", name="app_media_delete", methods={"POST"})
      */
     public function delete(Request $request, Media $media, MediaRepository $mediaRepository): Response
     {
-    //    if ($this->isCsrfTokenValid('delete'.$media->getId(), $request->request->get('_token'))) {
+       if ($this->isCsrfTokenValid('delete'.$media->getId(), $request->request->get('_token'))) {
             $mediaRepository->remove($media, true);
-    //    }
+       }
 
-        return $this->redirectToRoute('app_media_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_tricks_index', [], Response::HTTP_SEE_OTHER);
     }
-}
+}
